Add caching to the queue list portlet
On larger RTs with many queues, and users attached to many
tickets over time, the rights check to generate the appropriate
queue list for a user as been observed to take a significant
amount of time (20+ seconds on one system). Cache this list
much like the queue list in the "Create ticket" dropdown since
it changes infrequently from page load to page load.