WIP: Make the ticket timer window vertical resize work
Use 'display: table*' styled divs. This allows top and bottom edges of
table-cell styled divs to "stick" to each other, regardless of viewport
height.

Normally, we might use JavaScript to update the the viewport height as the
user resizes the window. But a new-ish CSS length measurement, "vh",
allows divs to change their heights in relation to the viewport height
automatically. This change also converts the ticket timer to use that
length measurement.

The vertical resizing component of this change works well. But
converting the ticket timer divs to use table display style required
breaking the existing styling, like borders and background color. Those
could probably be split up in to gutter-top, gutter-middle, and
gutter-bottom classes to apply to the 'table-header-row',
'table-row', and 'table-footer-row' styled divs.