Check role's direct group members for ticket watcher searches
For queries like "AdminCc.Name = 'foo'", the change is straightforward
as we can find all the user defined groups of user "foo" and search them
too.

For queries like "AdminCc.Name LIKE 'foo'" when there are multiple users
whose names contain "foo", it could logically work if we iterate all
those users and then search them using the above "=" strategy, which is
equivalent to "AdminCc.Name = 'foobar' OR AdminCc.Name = 'foobaz' OR
...", but I bet the performance will be bad as user list grows.

For criteria involving multiple users, previous code did a joined SQL of
CGMs and Users. As users might not be linked to role groups directly in
CGMs, "CGMs <-> Users" join is not enough, instead of checking if user
is a member of role group, we need another CGMs join to test if user is
a member of role group or its direct member groups.

Note that as role groups are treated as their own members in CGMs, we
can technially only test if user is a member of role group's direct
group members.

i.e. from

    CGMs(MemberId)  <->  Users(id)

to

    CGMs(MemberId)  <->  CGMs_2(GroupId)|CGMs_2(MemberId)  <->  Users(id)

For queries like "AdminCc.Name NOT LIKE 'foo'" when there are multiple
users whose names contain "foo", previous behavior is to return tickets
which have at least a member in AdminCc that doesn't have name matching
"foo". i.e. if a ticket's AdminCc contains users "foo" and "bar", then
"AdminCc.Name NOT LIKE 'foo'" will still return that ticket.  Because of
this behavior, there is no need to adjust code here and it will work as
before: if a ticket role contains user defined groups, then that ticket
will always return.

This was commented as "semi-working solution" and I'm afraid it'll be
quite hard if possible to properly correct it to make it totally
opposite of "LIKE". One workaround I can think of is to get user ids
first, then search them using "!=" instead, i.e. something like
"AdminCc.id != 2 AND AdminCc.id != 3 AND ...", which deserves another
branch, so it's not covered here.

See also 9444ce14