Check role's direct group members for ticket watcher group searches
For queries like "AdminCcGroup = 'foo'", we need to check if there is a
member in a ticket's AdminCc role that is in group "foo" or not.

Previously we did the joined search like:

    CGMs(MemberId)  <->  Users(id)  <->  CGMs_2(MemberId)

And check GroupId of CGMs_2 agaist the id of group "foo".

As users might not be linked to role groups directly in CGMs, the first
"CGMs <-> Users" join is not enough, we need another CGMs join to test
if user is a member of role group's direct member groups, i.e.

    CGMs(MemberId)  <->  CGMs_2(GroupId)|CGMs_2(MemberId)  <->  Users(id)  <->  CGMs_3(MemberId)

And then check GroupId of CGMs_3 agaist the id of group "foo".

See also 9444ce14 and 6974cf0c