Update layout to support wrapping subject
This commit makes changes to the positions, floats, padding, and
margins to allow the header h1 subject line to wrap.  The other
elements in the theme need to respond relative to the size of the
subject, so needed to be adjusted for position, padding, and margin
to compensate for the positioning change.