Update EmailAddress index to case insensitive for Pg
EmailAddress is case insensitive in RT(technically, name part could be
sensitive according to RFC, but it's confusing and rarely used in real
world). When we call RT::User::LoadByEmail, the WHERE part of generated
SQL in Pg is actually like:

    LOWER(EmailAddress) = LOWER(?)

Thus we need to adjust index accordingly, which is also consistent with
the one in Oracle.