Support to search user defined group names in watcher limit
To search tickets that have a specified group in a role, previously we
had to specify the group id like "AdminCc.id = 123", which is not
convenient. Here we support group names to improve this, e.g.
"AdminCc.Name = 'team foo'".

In case the name in question is shared by a user and a group(which is
weird and should be quite rare in real world), both will be respected.
I.e. we return tickets that have either the user or the group in the
role.