Fix main nav overlap on dark theme mobile
For smaller mobile browsers sizes like 375px wide, on dark theme,
the main nav overlaps the search box.  Raising the z-index of the
search box fixes the issue in the simplest way in the short term.

A better long term fix rather than relying on z-index is to rework
the element sizing and padding, and remove the blanket cascading
background color definition for dark theme.

Additionally, center the search box slightly.