Update prefs page to support single column layout
On mobile layout or smaller window sizes, the Prefs AboutMe page
was not correctly flowing to a single column layout, similar to
Admin/Users/Modify.  This commit fixes EditAboutMe to support the
single column at smaller sizes, and double column at larger.

Additionally, this commit fixes the col definition for the
signature box on the EditAboutMe page to match Admin/Users/Modify.